Adds carousel feature to checkbox.
This adds the carousel argument `Checkbox(Question)` class and
implements the feature in `Checkbox(BaseConsoleRender).process_input()`
just as previously done for the list. It adds the example script
`checkbox_carousel.py` as well as two acceptance and two integration
tests to verity moving around works. Last but not least it adds a
comment about the feature in the `README.md`.